+++ apps/libraries/Constant.php
@@ -23,7 +23,7 @@ class Constant
     const invalidUserPasskeyZset = 'Tracker:invalid_user_passkey:zset';
 
     // Tracker Use
-    const trackerInvalidInfoHashZset = 'Tracker:invalid_info_hash:zset';
+    const trackerInvalidInfoHashZset = 'Tracker:invalid_torrent_info_hash:zset';
     const trackerAllowedClientList = 'Tracker:allowed_client_list:string';
     const trackerAllowedClientExceptionList = 'Tracker:allowed_client_exception_list:string';
     const trackerValidClientZset = 'Tracker:valid_clients:zset';